requirements.txt
faster-whisper==1.0.2
grog==0.9
openai==1.30.1
google-generativeai==0.5.4
opencv-python
numpy<2.0.0
Pillow==10.3.0
PyAudio==0.2.14
SpeechRecognition==3.10.1
pyperclip==1.8.2
pytest==8.3.3
pytest-qt==4.2.0
flake8==6.1.0
black==23.12.1
whisper==1.1.10
sounddevice==0.4.6
PyQt6==6.6.0